Add files via upload
Graph showing the difference (delta) of the first vs. second half of the Boston Marathon against the total marathon time. Those who ran evenly should be at zero on the y-axis. Note that most data trend above this line. This data is uncleaned as of yet, as noted by the datapoints with an approx. y=ax+b relationship in the upper left quadrant of the graph. Participants to be focused on as having run "well" at 2016 Boston Marathon fall at or below the y=0 line.